savings id

+++ database/migrations/2025_10_27_084538_create_liquidation_addresses_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('bridge_liquidation_id'); 
             $table->string('chain');
-            $table->int('savings_id');
+            $table->unsignedBigInteger('savings_id');
             $table->string('address');
             $table->string('currency');
             $table->string('customer_id');
